+++ contracts/modules/STO/DummySTOFactory.sol
@@ -84,7 +84,7 @@ contract DummySTOFactory is ModuleFactory {
     }
 
     /**
-     * @notice Get the Instructions that helped to used the module
+     * @notice Returns the instructions associated with the module
      */
     function getInstructions() external view returns(string) {
         return "Dummy STO - you can mint tokens at will";